update package name to adjust to cookpad org
