--- app/src/androidTest/java/onesky/rejasupotaro/app/ExampleInstrumentedTest.kt
@@ -1,27 +1,0 @@
-package onesky.rejasupotaro.app
-
-import android.content.Context
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
-
-import org.junit.Test
-import org.junit.runner.RunWith
-
-import org.junit.Assert.*
-
-/**
- * Instrumentation test, which will execute on an Android device.
-
- * @see [Testing documentation](http://d.android.com/tools/testing)
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    @Throws(Exception::class)
-    fun useAppContext() {
-        // Context of the rejasupotaro.onesky.app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
-
-        assertEquals("rejasupotaro.onesky.app", appContext.packageName)
-    }
-}
