+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="onesky.rejasupotaro.app">
+    package="com.cookpad.android.onesky.app">
 
     <application
         android:allowBackup="true"
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="rejasupotaro.onesky.app.MainActivity">
+        <activity android:name="com.cookpad.android.onesky.app.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
