+++ app/src/main/java/com/cookpad/android/onesky/app/MainActivity.kt
@@ -1,0 +1,13 @@
+package com.cookpad.android.onesky.app
+
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+import onesky.rejasupotaro.app.R
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
