--- app/src/main/java/rejasupotaro/onesky/app/MainActivity.kt
@@ -1,14 +1,0 @@
-package rejasupotaro.onesky.app
-
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-
-import onesky.rejasupotaro.app.R
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
