--- plugin/src/main/java/rejasupotaro/onesky/plugin/Files.kt
@@ -1,10 +1,0 @@
-package rejasupotaro.onesky.plugin
-
-internal fun valuesDirNameFromLocale(locale: String): String {
-    return "values-${locale.replace("id", "in").replace("-", "-r")}"
-}
-
-internal fun localeFromValuesDirName(dirName: String): String {
-    return dirName.replace("values-", "").replace("in", "id").replace("-r", "-")
-}
-
