--- plugin/src/main/java/rejasupotaro/onesky/plugin/tasks/UploadTranslationTask.kt
@@ -1,29 +1,0 @@
-package rejasupotaro.onesky.plugin.tasks
-
-import com.github.kittinunf.result.Result
-import org.gradle.api.tasks.TaskAction
-import java.io.File
-
-open class UploadTranslationTask : OneskyTask() {
-    init {
-        group = "Translation"
-        description = "Upload the default translation file (values/strings.xml)"
-    }
-
-    @TaskAction
-    fun uploadTranslation() {
-        val file = File("${project.projectDir.absolutePath}/src/main/res/values/strings.xml")
-        print("Uploading ${file.absolutePath} ... ")
-        val result = oneskyClient.upload(file)
-        when (result) {
-            is Result.Success -> {
-                println("Done!")
-            }
-            is Result.Failure -> {
-                println("Failed!")
-                throw result.error
-            }
-        }
-    }
-}
-
