+++ plugin/src/main/resources/META-INF/gradle-plugins/com.cookpad.android.onesky.plugin.properties
@@ -1,0 +1,1 @@
+implementation-class=com.cookpad.android.onesky.plugin.OneskyPlugin
