--- plugin/src/main/resources/META-INF/gradle-plugins/rejasupotaro.onesky.plugin.properties
@@ -1,1 +1,0 @@
-implementation-class=rejasupotaro.onesky.plugin.OneskyPlugin
